Refactor GUI into modular views
